chore(metrics-modal): move to /packages
