+++ packages/administration/src/metrics/index.ts
@@ -1,0 +1,2 @@
+export { MetricModule } from './metric.module';
+export { JhiMetricsMonitoringComponent } from './metrics.component';
