+++ packages/administration/src/metrics/metrics-modal.component.html
