+++ packages/administration/src/metrics/metrics.component.html
