+++ packages/administration/src/metrics/metrics.component.scss
