+++ packages/administration/src/metrics/metrics.component.ts
