--- src/app/admin/metrics/index.ts
@@ -1,1 +1,0 @@
-export { MetricModule } from './metric.module';
